refactor(upload): mutualisation du parsing de la progression d'intégration
